add command line options `--crl`, `--cr1`, `--cr2` for `two_mode_reduction` 	modified:   src/cdb/db.rs 	modified:   src/cdb/mod.rs 	modified:   src/config.rs 	modified:   src/solver/search.rs

+++ src/cdb/mod.rs
@@ -308,7 +308,7 @@ pub enum ReductionType {
     /// weight by Literal Block Distance over the added clauses
     LBDonADD(usize),
     /// weight by Literal Block Distance over all learnt clauses
-    LBDonALL(f64, f64),
+    LBDonALL(u16, f64),
 }
 
 pub mod property {
